Simplification components
LEs composants simple ont étés déplacés directement dans le src
