--- src/components/APropos/APropos.js
@@ -1,20 +1,0 @@
-import React from "react";
-
-const APropos = () => {
-    return (
-        <>
-        <h2>Wordlet</h2>
-			<p>
-				<br/>Projet de BIHANNIC Lucas, RICHARD Damien et TRONEL Ludovic.
-				<br/> Tous 3 étudiants en licence Informatique à l'Université de la Nouvelle-Calédonie.
-				<br/>Wordlet, de la compression de Word Wallet, littéralement traduit par "le portefeuille de mots", 
-				est une application permettant l'introduction au monde des blockchain par le biais de NFT (Token Non Fongible) 
-				sous forme de mots, qui peuvent être achetés, vendus, échangés où joués et gagnés par le biais de différents 
-				jeux. Ces derniers peuvent être codés par les utilisateurs à partir de code open source, les meilleurs d'entre 
-				eux sont votés par les joueurs et les plus populaires seront implémentés par les créateurs de l'application. 
-            </p>
-        </>
-    )
-}
-
-export default APropos
