+++ src/components/Accueil.js
@@ -1,13 +1,13 @@
 import React, {useState, useEffect} from 'react'
 import * as fcl from "@onflow/fcl"
 
-import { useAuth } from '../../providers/AuthProvider';
-import { useUser } from '../../providers/UserProvider';
+import { useAuth } from '../providers/AuthProvider';
+import { useUser } from '../providers/UserProvider';
 
 const Accueil = () => {
     const { user, loggedIn, logOut } = useAuth()
     return(
-        <>
+        <section id="accueil">
             <h1 id="wordlet-title">Wordlet</h1>
             {loggedIn?
                 <div id="auth-cluster">
@@ -17,7 +17,11 @@ const Accueil = () => {
             :
                 <button className="btn-primary" id="auth-cluster" onClick={fcl.authenticate}>Connexion</button>
             }
-        </>
+            <img src="etoiles.png" id="stars" alt=""/>
+            <img src="lune.png" id="moon" alt=""/>
+            <img src="montagne_derriere.png" id="mountains_behind" alt=""/>
+
+        </section>
     )
 }
